代码/requirements.txt
pyyaml
opencv-python
numpy
werkzeug
jsonrpc-py
RPi.GPIO
smbus2
rpi_ws281x
matplotlib
torch
ultralytics
pandas
Pillow
